make callback signer test pass
